+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- [ENHANCEMENT] Expose metrics in the prometheus format - [#2](//github.com/jeffersongirao/source-ranges-controller/issues/2)
 - [ENHANCEMENT] Added ability to restrict to a namespace - [#1](//github.com/jeffersongirao/source-ranges-controller/issues/1)
 
 ## [0.0.1] / 2018-04-24